+++ heron/packer/src/python/BUILD
@@ -3,12 +3,12 @@ package(default_visibility = ["//visibility:public"])
 load("/tools/rules/pex_rules", "pex_library", "pex_binary")
 
 pex_library(
-    name = "heron-package-lib",
+    name = "heron-packer-lib",
     srcs = [
       "store/metastore.py",
       "store/blobstore.py",
       "common/utils.py",
-      "common/package.py",
+      "common/packer.py",
       "common/constants.py",
       "common/parser.py",
     ],
@@ -19,11 +19,11 @@ pex_library(
 )
 
 pex_binary(
-    name = "heron-package",
+    name = "heron-packer",
     srcs = [
         "main.py",
     ],
     deps = [
-        ":heron-package-lib",
+        ":heron-packer-lib",
     ]
 )
